Created Xcode Project with Login and DJ Table
Deleted storyboard, set the window's rootview to be a navigation controller with the DJ TableView Controller as view. Added Logout button to take you back to the login controller. Login controller has login button, username and password views. We need to implement a dual button for DJ Enter and Guest enter